facts!

+++ project/src/backend/sdl/SDLApplication.h
@@ -62,8 +62,8 @@ namespace lime {
 			bool active;
 			ApplicationEvent applicationEvent;
 			ClipboardEvent clipboardEvent;
-			double currentUpdate;
-			double framePeriod;
+			int64_t currentUpdate;
+			int64_t framePeriod;
 			Uint32 initFlags;
 			DropEvent dropEvent;
 			GamepadEvent gamepadEvent;
